pyproject.toml
[tool.poetry]
name = "tier2-backend-python-fastapi-sentry-pydantic"
version = "0.1.0"
description = "Tier 2 Backend - FastAPI with Sentry and Pydantic"
authors = ["Code Campfire <info@codecampfire.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
python-dotenv = "^1.0.0"
sentry-sdk = "^1.40.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
httpx = "^0.26.0"
pytest-cov = "^4.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 88
target-version = ['py39']
exclude = '''
/(
    \.git
  | \.venv
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 88 